updated Chip documentation
